refactor: Split node queries in another file
This also makes the file lighter
