+++ src/queries/queries.js
@@ -15,9 +15,7 @@ import {
   CCO2_SETTINGS_DOCTYPE,
   CONTACTS_DOCTYPE,
   FILES_DOCTYPE,
-  SETTINGS_DOCTYPE,
-  KONNECTORS_DOCTYPE,
-  CONTACTS_GROUPS_DOCTYPE
+  SETTINGS_DOCTYPE
 } from 'src/doctypes'
 
 import CozyClient, { Q } from 'cozy-client'
@@ -268,25 +266,6 @@ export const buildTimeseriesQueryByDateAndAccountId = ({
   }
 }
 
-// Note there is no need for fetchPolicy here, as this query is only used in node service
-export const buildTimeseriesWithoutAggregation = ({ limit = 1000 } = {}) => ({
-  definition: Q(GEOJSON_DOCTYPE)
-    .include(['startPlaceContact', 'endPlaceContact'])
-    .where({
-      startDate: {
-        $gt: null
-      }
-    })
-    .partialIndex({
-      aggregation: {
-        $exists: false
-      }
-    })
-    .indexFields(['startDate'])
-    .sortBy([{ startDate: 'desc' }])
-    .limitBy(limit)
-})
-
 export const buildOneYearOldTimeseriesWithAggregationForAllAccounts = () => {
   const dateOneYearAgoFromNow = startOfMonth(subYears(new Date(), 1))
 
@@ -445,147 +424,7 @@ export const queryContactByDocId = async (client, docId) => {
   return res?.data
 }
 
-export const queryAccountByDocId = async (client, docId) => {
-  const res = await client.query(Q(ACCOUNTS_DOCTYPE).getById(docId))
-  return res?.data
-}
-
-// ---------- Node.js queries
-// accounId from buildAccountQuery function
-export const buildOldestTimeseriesQueryByAccountId = accountId => ({
-  definition: Q(GEOJSON_DOCTYPE)
-    .where({
-      'cozyMetadata.sourceAccount': accountId
-    })
-    .indexFields(['cozyMetadata.sourceAccount', 'startDate'])
-    .sortBy([{ 'cozyMetadata.sourceAccount': 'asc' }, { startDate: 'asc' }])
-    .limitBy(1)
-})
-
-export const buildNewestRecurringTimeseriesQueryForAllAccounts = () => ({
-  definition: Q(GEOJSON_DOCTYPE)
-    .where({})
-    .partialIndex({
-      'aggregation.recurring': true
-    })
-    .indexFields(['startDate'])
-    .sortBy([{ 'cozyMetadata.sourceAccount': 'desc' }, { startDate: 'desc' }])
-    .limitBy(1)
-})
-export const buildNewestRecurringTimeseriesQuery = ({ accountId }) => {
-  return {
-    definition: Q(GEOJSON_DOCTYPE)
-      .where({
-        'cozyMetadata.sourceAccount': accountId
-      })
-      .partialIndex({
-        'aggregation.recurring': true
-      })
-      .indexFields(['cozyMetadata.sourceAccount', 'startDate'])
-      .sortBy([{ 'cozyMetadata.sourceAccount': 'desc' }, { startDate: 'desc' }])
-      .limitBy(1)
-  }
-}
-
-// accounId from timeseries (timeseries.cozyMetadata.sourceAccount)
-export const buildRecurringTimeseriesByStartAndEndPointRange = ({
-  accountId,
-  minLatStart,
-  maxLatStart,
-  minLonStart,
-  maxLonStart,
-  minLonEnd,
-  maxLonEnd,
-  minLatEnd,
-  maxLatEnd,
-  limit = 1000
-}) => {
-  return {
-    definition: Q(GEOJSON_DOCTYPE)
-      .where({
-        'cozyMetadata.sourceAccount': accountId,
-        'aggregation.coordinates.startPoint.lon': {
-          $gte: minLonStart,
-          $lte: maxLonStart
-        },
-        'aggregation.coordinates.startPoint.lat': {
-          $gte: minLatStart,
-          $lte: maxLatStart
-        },
-        'aggregation.coordinates.endPoint.lon': {
-          $gte: minLonEnd,
-          $lte: maxLonEnd
-        },
-        'aggregation.coordinates.endPoint.lat': {
-          $gte: minLatEnd,
-          $lte: maxLatEnd
-        }
-      })
-      .partialIndex({
-        'aggregation.recurring': true
-      })
-      .indexFields([
-        'cozyMetadata.sourceAccount',
-        'aggregation.coordinates.startPoint.lon',
-        'aggregation.coordinates.startPoint.lat',
-        'aggregation.coordinates.endPoint.lon',
-        'aggregation.coordinates.endPoint.lat'
-      ])
-      .limitBy(limit)
-  }
-}
-
-// accounId from buildLastCreatedServiceAccountQuery && buildAccountByToken functions
-export const buildTimeseriesByDateRange = ({
-  firstDate,
-  lastDate,
-  accountId,
-  withRelationships = false,
-  limit = 100
-}) => {
-  const queryDef = Q(GEOJSON_DOCTYPE)
-    .where({
-      'cozyMetadata.sourceAccount': accountId,
-      startDate: {
-        $gte: firstDate,
-        $lte: lastDate
-      }
-    })
-    .indexFields(['cozyMetadata.sourceAccount', 'startDate'])
-    .sortBy([{ 'cozyMetadata.sourceAccount': 'desc' }, { startDate: 'desc' }])
-    .limitBy(limit)
-  if (withRelationships) {
-    // FIXME: .include method appears to do nothing. It should be investigated
-    queryDef.includes = ['startPlaceContact', 'endPlaceContact']
-  }
-  return {
-    definition: queryDef,
-    options: {
-      as: `${GEOJSON_DOCTYPE}/dateRange/${firstDate}-${lastDate}/accountId/${accountId}`
-    }
-  }
-}
-
 // other doctypes -------------
-
-export const buildContactsWithGeoCoordinates = ({ limit = 1000 } = {}) => {
-  return {
-    definition: Q(CONTACTS_DOCTYPE)
-      .partialIndex({
-        address: {
-          $elemMatch: {
-            'geo.geo': {
-              $exists: true
-            }
-          }
-        }
-      })
-      .limitBy(limit),
-    options: {
-      as: `${CONTACTS_DOCTYPE}/withGeoCoordinates`
-    }
-  }
-}
 
 export const buildAccountQuery = ({
   limit = 100,
@@ -630,25 +469,6 @@ export const buildLastCreatedServiceAccountQuery = () => {
     definition: queryDef,
     options: {
       as: `${ACCOUNTS_DOCTYPE}/account_type/openpath-with-token`,
-      fetchPolicy: CozyClient.fetchPolicies.olderThan(older30s)
-    }
-  }
-}
-
-export const buildAccountByToken = ({ token }) => {
-  const queryDef = Q(ACCOUNTS_DOCTYPE)
-    .where({
-      token: token
-    })
-    .partialIndex({
-      account_type: 'openpath'
-    })
-    .indexFields(['token'])
-    .limitBy(1)
-  return {
-    definition: queryDef,
-    options: {
-      as: `${ACCOUNTS_DOCTYPE}/account_type/openpath/token/${token}`,
       fetchPolicy: CozyClient.fetchPolicies.olderThan(older30s)
     }
   }
@@ -705,14 +525,6 @@ export const buildFileQueryById = fileId => ({
   }
 })
 
-export const buildOpenPathKonnectorQuery = () => ({
-  definition: Q('io.cozy.konnectors').getById('openpath'),
-  options: {
-    as: `${KONNECTORS_DOCTYPE}/openpath`,
-    fetchPolicy: CozyClient.fetchPolicies.olderThan(older30s)
-  }
-})
-
 export const buildAccountQueryByLogin = login => ({
   definition: Q(ACCOUNTS_DOCTYPE)
     .where({
@@ -729,16 +541,3 @@ export const buildAccountQueryByLogin = login => ({
     as: `${ACCOUNTS_DOCTYPE}/account_type/login/${login}`
   }
 })
-
-export const buildContactsGroupByName = ({ groupName }) => {
-  return {
-    definition: Q(CONTACTS_GROUPS_DOCTYPE)
-      .where({
-        name: groupName
-      })
-      .limitBy(1),
-    options: {
-      as: `${CONTACTS_GROUPS_DOCTYPE}/byName/${groupName}`
-    }
-  }
-}
